Switched Apostle to be a static library instead of a dynamic library

+++ Apostle/src/Apostle/ImGui/ImGuiLayer.cpp
@@ -11,7 +11,6 @@
 
 #include <GLFW/glfw3.h>
 
-
 namespace Apostle {
 
     ImGuiLayer::ImGuiLayer()
@@ -84,7 +83,7 @@ namespace Apostle {
     {
         ImGuiIO& io = ImGui::GetIO();
         Application& app = Application::Get();
-        io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+        io.DisplaySize = ImVec2((float)app.GetWindow().GetWidth(), (float)app.GetWindow().GetHeight());
         
         // Rendering
         ImGui::Render();
